rename to time control bar
