--- frontend/src/metabase/embedding/mcp/QueryExplorerBar/index.ts
@@ -1,8 +1,0 @@
-export { QueryExplorerBar } from "./QueryExplorerBar";
-export type {
-  QueryExplorerBarProps,
-  QueryExplorerBarChartType,
-  TimeRangeConfig,
-  TimeGranularityConfig,
-  TimeGranularityItem,
-} from "./QueryExplorerBar";
