+++ frontend/src/metabase/embedding/mcp/TimeControlBar/ChartTypePicker.module.css
