+++ frontend/src/metabase/embedding/mcp/TimeControlBar/ChartTypePicker.tsx
