+++ frontend/src/metabase/embedding/mcp/TimeControlBar/TimeGranularityControl.tsx
