+++ frontend/src/metabase/embedding/mcp/TimeControlBar/index.ts
@@ -1,0 +1,7 @@
+export { TimeControlBar } from "./TimeControlBar";
+export type {
+  TimeControlBarProps,
+  TimeRangeConfig,
+  TimeGranularityConfig,
+  TimeGranularityItem,
+} from "./TimeControlBar";
